LayoutTests/fast/events/no-scroll-on-input-text-selection-expected.txt

Verify selecting text does not cause any scrolling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ta.selectionStart < ta.selectionEnd is true
PASS pageXOffset is 0
PASS pageYOffset is originalPageYOffset
PASS document.scrollingElement.scrollLeft is 0
PASS document.scrollingElement.scrollTop is originalBodyScrollTop
PASS pageYOffset is document.scrollingElement.scrollTop
PASS pageYOffset > 2000 is true
PASS successfullyParsed is true

TEST COMPLETE
Test for bug 12331.

To test manually, scroll down until you find the input element. Select some text. If the bug occurs, the page will scroll back up to the top.


If the bug does not occur, you'll be left down here at the bottom.
